fix(prepare_test_table): wait for agreement while changing schema
C-S thread creating keyspaces in this method may fail to write data
and freeze if the nodes do not get to schema agreement quickly enough
(due to Scylla problem with schema management).
To avoid it we will wait for schema agreement while creating a keyspace
and a table. We will start using `IF NOT EXISTS` so we won't need the
additional check of ks and table presence anymore.
Also, we will stop using C-S thread and switch to Python driver and CQL
to populate the created tables.

This workaround should fix https://github.com/scylladb/scylla-cluster-tests/issues/5413